+++ rs/nervous_system/timestamp/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "ic-nervous-system-timestamp"
+version = "0.0.1"
+edition = "2021"
+
+[dependencies]
+time = { workspace = true }
